requirements.txt
pillow>=10.0.0
mss>=9.0.0
pystray>=0.19.0
